requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
replicate>=0.21.0
python-dotenv>=1.0.0


python-telegram-bot==20.7
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0